chore: Move internal components to the correct folder
