+++ src/internal/components/drag-handle/styles.scss
