+++ src/internal/components/handle/styles.scss
@@ -3,9 +3,9 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-@use '../../internal/styles' as styles;
-@use '../../internal/hooks/focus-visible' as focus-visible;
-@use '../../internal/styles/tokens' as awsui;
+@use '../../styles' as styles;
+@use '../../hooks/focus-visible' as focus-visible;
+@use '../../styles/tokens' as awsui;
 
 .handle {
   appearance: none;
