+++ src/test-utils/dom/collection-preferences/content-display-preference.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import styles from '../../../collection-preferences/styles.selectors.js';
-import dragHandleStyles from '../../../internal/drag-handle/styles.selectors.js';
+import dragHandleStyles from '../../../internal/components/drag-handle/styles.selectors.js';
 import { ComponentWrapper, ElementWrapper } from '@cloudscape-design/test-utils-core/dom';
 import ToggleWrapper from '../toggle';
